Moving Tab Manager and WKUIDelegate to extension Adding long-press actions for recently closed tabs action clicks
