+++ Client/Frontend/Browser/BrowserViewController.swift
@@ -95,7 +95,7 @@ public class BrowserViewController: UIViewController {
   }()
   
   var readerModeBar: ReaderModeBarView?
-  private(set) var readerModeCache: ReaderModeCache
+  var readerModeCache: ReaderModeCache
   
   private let statusBarOverlay: UIView = {
     // Temporary work around for covering the non-clipped web view content
@@ -171,7 +171,7 @@ public class BrowserViewController: UIViewController {
   }
   private var toolbarTopConstraint: Constraint?
   private var toolbarBottomConstraint: Constraint?
-  private var toolbarVisibilityCancellable: AnyCancellable?
+  var toolbarVisibilityCancellable: AnyCancellable?
 
   var keyboardState: KeyboardState?
 
@@ -1771,7 +1771,7 @@ public class BrowserViewController: UIViewController {
   }
 
   /// Updates the URL bar security, text and button states.
-  fileprivate func updateURLBar() {
+  func updateURLBar() {
     guard let tab = tabManager.selectedTab else { return }
 
     updateRewardsButtonState()
@@ -2251,7 +2251,7 @@ public class BrowserViewController: UIViewController {
     }
   }
   
-  private func handleToolbarVisibilityStateChange(
+  func handleToolbarVisibilityStateChange(
     _ state: ToolbarVisibilityViewModel.ToolbarState,
     progress: CGFloat?
   ) {
@@ -2663,591 +2663,6 @@ extension BrowserViewController: SearchViewControllerDelegate {
       return false
     }
     return true
-  }
-}
-
-extension BrowserViewController: TabManagerDelegate {
-  func tabManager(_ tabManager: TabManager, didSelectedTabChange selected: Tab?, previous: Tab?) {
-    // Remove the old accessibilityLabel. Since this webview shouldn't be visible, it doesn't need it
-    // and having multiple views with the same label confuses tests.
-    if let wv = previous?.webView {
-      toolbarVisibilityViewModel.endScrollViewObservation(wv.scrollView)
-      
-      wv.endEditing(true)
-      wv.accessibilityLabel = nil
-      wv.accessibilityElementsHidden = true
-      wv.accessibilityIdentifier = nil
-      wv.removeFromSuperview()
-    }
-
-    toolbar?.setSearchButtonState(url: selected?.url)
-    if let tab = selected, let webView = tab.webView {
-      toolbarVisibilityViewModel.beginObservingScrollView(webView.scrollView)
-      toolbarVisibilityCancellable = toolbarVisibilityViewModel.objectWillChange
-        .receive(on: DispatchQueue.main)
-        .sink(receiveValue: { [weak self] in
-          guard let self = self else { return }
-          let (state, progress) = (self.toolbarVisibilityViewModel.toolbarState,
-                                   self.toolbarVisibilityViewModel.interactiveTransitionProgress)
-          self.handleToolbarVisibilityStateChange(state, progress: progress)
-        })
-      
-      updateURLBar()
-
-      if let url = tab.url, !InternalURL.isValid(url: url) {
-        let previousEstimatedProgress = previous?.webView?.estimatedProgress ?? 1.0
-        let selectedEstimatedProgress = webView.estimatedProgress
-
-        // Progress should be updated only if there's a difference between tabs.
-        // Otherwise we do nothing, so switching between fully loaded tabs won't show the animation.
-        if previousEstimatedProgress != selectedEstimatedProgress {
-          topToolbar.updateProgressBar(Float(selectedEstimatedProgress))
-        }
-      } else {
-        topToolbar.hideProgressBar()
-      }
-
-      readerModeCache = ReaderModeScriptHandler.cache(for: tab)
-      ReaderModeHandler.readerModeCache = readerModeCache
-
-      webViewContainer.addSubview(webView)
-      webView.snp.remakeConstraints { make in
-        make.left.right.top.bottom.equalTo(self.webViewContainer)
-      }
-
-      // This is a terrible workaround for a bad iOS 12 bug where PDF
-      // content disappears any time the view controller changes (i.e.
-      // the user taps on the tabs tray). It seems the only way to get
-      // the PDF to redraw is to either reload it or revisit it from
-      // back/forward list. To try and avoid hitting the network again
-      // for the same PDF, we revisit the current back/forward item and
-      // restore the previous scrollview zoom scale and content offset
-      // after a short 100ms delay. *facepalm*
-      //
-      // https://bugzilla.mozilla.org/show_bug.cgi?id=1516524
-      if tab.mimeType == MIMEType.PDF {
-        let previousZoomScale = webView.scrollView.zoomScale
-        let previousContentOffset = webView.scrollView.contentOffset
-
-        if let currentItem = webView.backForwardList.currentItem {
-          webView.go(to: currentItem)
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + .milliseconds(100)) {
-          webView.scrollView.setZoomScale(previousZoomScale, animated: false)
-          webView.scrollView.setContentOffset(previousContentOffset, animated: false)
-        }
-      }
-
-      webView.accessibilityLabel = Strings.webContentAccessibilityLabel
-      webView.accessibilityIdentifier = "contentView"
-      webView.accessibilityElementsHidden = false
-
-      if webView.url == nil {
-        // The web view can go gray if it was zombified due to memory pressure.
-        // When this happens, the URL is nil, so try restoring the page upon selection.
-        tab.reload()
-      }
-    }
-
-    updateToolbarUsingTabManager(tabManager)
-
-    removeAllBars()
-    if let bars = selected?.bars {
-      for bar in bars {
-        showBar(bar, animated: true)
-      }
-    }
-
-    updateFindInPageVisibility(visible: false, tab: previous)
-    displayPageZoom(visible: false)
-    updateTabsBarVisibility()
-    selected?.updatePullToRefreshVisibility()
-
-    topToolbar.locationView.loading = selected?.loading ?? false
-    navigationToolbar.updateBackStatus(selected?.canGoBack ?? false)
-    navigationToolbar.updateForwardStatus(selected?.canGoForward ?? false)
-
-    let shouldShowPlaylistURLBarButton = selected?.url?.isPlaylistSupportedSiteURL == true
-
-    if let readerMode = selected?.getContentScript(name: ReaderModeScriptHandler.scriptName) as? ReaderModeScriptHandler,
-      !shouldShowPlaylistURLBarButton {
-      topToolbar.updateReaderModeState(readerMode.state)
-      if readerMode.state == .active {
-        showReaderModeBar(animated: false)
-      } else {
-        hideReaderModeBar(animated: false)
-      }
-
-      updatePlaylistURLBar(tab: selected, state: selected?.playlistItemState ?? .none, item: selected?.playlistItem)
-    } else {
-      topToolbar.updateReaderModeState(ReaderModeState.unavailable)
-    }
-
-    updateInContentHomePanel(selected?.url as URL?)
-
-    notificationsPresenter.removeNotification(with: WalletNotification.Constant.id)
-    WalletProviderPermissionRequestsManager.shared.cancelAllPendingRequests(for: [.eth, .sol])
-    WalletProviderAccountCreationRequestManager.shared.cancelAllPendingRequests(coins: [.eth, .sol])
-    updateURLBarWalletButton()
-  }
-
-  func tabManager(_ tabManager: TabManager, willAddTab tab: Tab) {
-  }
-
-  func tabManager(_ tabManager: TabManager, didAddTab tab: Tab) {
-    // If we are restoring tabs then we update the count once at the end
-    if !tabManager.isRestoring {
-      updateToolbarUsingTabManager(tabManager)
-    }
-    tab.tabDelegate = self
-    tab.walletKeyringService = BraveWallet.KeyringServiceFactory.get(privateMode: tab.isPrivate)
-    updateTabsBarVisibility()
-  }
-
-  func tabManager(_ tabManager: TabManager, willRemoveTab tab: Tab) {
-    tab.webView?.removeFromSuperview()
-  }
-
-  func tabManager(_ tabManager: TabManager, didRemoveTab tab: Tab) {
-    updateToolbarUsingTabManager(tabManager)
-    // tabDelegate is a weak ref (and the tab's webView may not be destroyed yet)
-    // so we don't expcitly unset it.
-    topToolbar.leaveOverlayMode(didCancel: true)
-    updateTabsBarVisibility()
-
-    if !PrivateBrowsingManager.shared.isPrivateBrowsing {
-      rewards.reportTabClosed(tabId: Int(tab.rewardsId))
-    }
-  }
-
-  func tabManagerDidAddTabs(_ tabManager: TabManager) {
-    updateToolbarUsingTabManager(tabManager)
-  }
-
-  func tabManagerDidRestoreTabs(_ tabManager: TabManager) {
-    updateToolbarUsingTabManager(tabManager)
-  }
-
-  func show(toast: Toast, afterWaiting delay: DispatchTimeInterval = SimpleToastUX.toastDelayBefore, duration: DispatchTimeInterval? = SimpleToastUX.toastDismissAfter) {
-    if let downloadToast = toast as? DownloadToast {
-      self.downloadToast = downloadToast
-    }
-
-    // If BVC isnt visible hold on to this toast until viewDidAppear
-    if self.view.window == nil {
-      self.pendingToast = toast
-      return
-    }
-
-    toast.showToast(
-      viewController: self, delay: delay, duration: duration,
-      makeConstraints: { make in
-        make.left.right.equalTo(self.view)
-        make.bottom.equalTo(self.webViewContainer)
-      })
-  }
-
-  func tabManagerDidRemoveAllTabs(_ tabManager: TabManager, toast: ButtonToast?) {
-    guard let toast = toast, !PrivateBrowsingManager.shared.isPrivateBrowsing else {
-      return
-    }
-    show(toast: toast, afterWaiting: ButtonToastUX.toastDelay)
-  }
-
-  func updateToolbarUsingTabManager(_ tabManager: TabManager) {
-    // Update Tab Count on Tab-Tray Button
-    let count = tabManager.tabsForCurrentMode.count
-    toolbar?.updateTabCount(count)
-    topToolbar.updateTabCount(count)
-
-    // Update Actions for Tab-Tray Button
-    var newTabMenuChildren: [UIAction] = []
-    var addTabMenuChildren: [UIAction] = []
-
-    if !PrivateBrowsingManager.shared.isPrivateBrowsing {
-      let openNewPrivateTab = UIAction(
-        title: Strings.newPrivateTabTitle,
-        image: UIImage(systemName: "plus.square.fill.on.square.fill"),
-        handler: UIAction.deferredActionHandler { [unowned self] _ in
-          self.openBlankNewTab(attemptLocationFieldFocus: false, isPrivate: true)
-        })
-
-      if (UIDevice.current.userInterfaceIdiom == .pad && tabsBar.view.isHidden == true) || (UIDevice.current.userInterfaceIdiom == .phone && toolbar == nil) {
-        newTabMenuChildren.append(openNewPrivateTab)
-      }
-
-      addTabMenuChildren.append(openNewPrivateTab)
-    }
-
-    let openNewTab = UIAction(
-      title: PrivateBrowsingManager.shared.isPrivateBrowsing ? Strings.newPrivateTabTitle : Strings.newTabTitle,
-      image: PrivateBrowsingManager.shared.isPrivateBrowsing ? UIImage(systemName: "plus.square.fill.on.square.fill") : UIImage(systemName: "plus.square.on.square"),
-      handler: UIAction.deferredActionHandler { [unowned self] _ in
-        self.openBlankNewTab(attemptLocationFieldFocus: false, isPrivate: PrivateBrowsingManager.shared.isPrivateBrowsing)
-      })
-
-    if (UIDevice.current.userInterfaceIdiom == .pad && tabsBar.view.isHidden) || (UIDevice.current.userInterfaceIdiom == .phone && toolbar == nil) {
-      newTabMenuChildren.append(openNewTab)
-    }
-    addTabMenuChildren.append(openNewTab)
-
-    var bookmarkMenuChildren: [UIAction] = []
-    
-    let containsWebPage = tabManager.selectedTab?.containsWebPage == true
-
-    if tabManager.openedWebsitesCount > 0, containsWebPage {
-      let bookmarkActiveTab = UIAction(
-        title: Strings.addToMenuItem,
-        image: UIImage(systemName: "book"),
-        handler: UIAction.deferredActionHandler { [unowned self] _ in
-          self.openAddBookmark()
-        })
-
-      bookmarkMenuChildren.append(bookmarkActiveTab)
-    }
-    
-    if tabManager.tabsForCurrentMode.count > 1 {
-      let bookmarkAllTabs = UIAction(
-        title: String.localizedStringWithFormat(Strings.bookmarkAllTabsTitle, tabManager.tabsForCurrentMode.count),
-        image: UIImage(systemName: "book"),
-        handler: UIAction.deferredActionHandler { [unowned self] _ in
-          let mode = BookmarkEditMode.addFolderUsingTabs(title: Strings.savedTabsFolderTitle, tabList: tabManager.tabsForCurrentMode)
-          let addBookMarkController = AddEditBookmarkTableViewController(bookmarkManager: bookmarkManager, mode: mode)
-
-          presentSettingsNavigation(with: addBookMarkController, cancelEnabled: true)
-        })
-
-      bookmarkMenuChildren.append(bookmarkAllTabs)
-    }
-    
-    var duplicateTabMenuChildren: [UIAction] = []
-
-    if containsWebPage, let selectedTab = tabManager.selectedTab, let url = selectedTab.fetchedURL {
-      let duplicateActiveTab = UIAction(
-        title: Strings.duplicateActiveTab,
-        image: UIImage(systemName: "plus.square.on.square"),
-        handler: UIAction.deferredActionHandler { [weak selectedTab] _ in
-          guard let selectedTab = selectedTab else { return }
-          
-          tabManager.addTabAndSelect(
-               URLRequest(url: url),
-               afterTab: selectedTab,
-               isPrivate: selectedTab.isPrivate
-           )
-        })
-
-      duplicateTabMenuChildren.append(duplicateActiveTab)
-    }
-
-    var closeTabMenuChildren: [UIAction] = []
-
-    let closeActiveTab = UIAction(
-      title: String(format: Strings.closeTabTitle),
-      image: UIImage(systemName: "xmark"),
-      attributes: .destructive,
-      handler: UIAction.deferredActionHandler { [unowned self] _ in
-        if let tab = tabManager.selectedTab {
-          if topToolbar.locationView.readerModeState == .active {
-            hideReaderModeBar(animated: false)
-          }
-          
-          tabManager.removeTab(tab)
-        }
-      })
-
-    closeTabMenuChildren.append(closeActiveTab)
-
-    if tabManager.tabsForCurrentMode.count > 1 {
-      let closeAllTabs = UIAction(
-        title: String(format: Strings.closeAllTabsTitle, tabManager.tabsForCurrentMode.count),
-        image: UIImage(systemName: "xmark"),
-        attributes: .destructive,
-        handler: UIAction.deferredActionHandler { [unowned self] _ in
-
-          let alert = UIAlertController(title: nil, message: Strings.closeAllTabsPrompt, preferredStyle: .actionSheet)
-          let cancelAction = UIAlertAction(title: Strings.CancelString, style: .cancel)
-          let closedTabsTitle = String(format: Strings.closeAllTabsTitle, tabManager.tabsForCurrentMode.count)
-          let closeAllAction = UIAlertAction(title: closedTabsTitle, style: .destructive) { _ in
-            tabManager.removeAllForCurrentMode()
-          }
-          alert.addAction(closeAllAction)
-          alert.addAction(cancelAction)
-
-          if let popoverPresentation = alert.popoverPresentationController {
-            let tabsButton = toolbar?.tabsButton ?? topToolbar.tabsButton
-            popoverPresentation.sourceView = tabsButton
-            popoverPresentation.sourceRect =
-              .init(x: tabsButton.frame.width / 2, y: tabsButton.frame.height, width: 1, height: 1)
-          }
-
-          self.present(alert, animated: true)
-        })
-
-      closeTabMenuChildren.append(closeAllTabs)
-    }
-
-    let newTabMenu = UIMenu(title: "", options: .displayInline, children: newTabMenuChildren)
-    let addTabMenu = UIMenu(title: "", options: .displayInline, children: addTabMenuChildren)
-    let bookmarkMenu = UIMenu(title: "", options: .displayInline, children: bookmarkMenuChildren)
-    let duplicateTabMenu = UIMenu(title: "", options: .displayInline, children: duplicateTabMenuChildren)
-    let closeTabMenu = UIMenu(title: "", options: .displayInline, children: closeTabMenuChildren)
-    
-    let tabButtonMenuActionList = [closeTabMenu, duplicateTabMenu, bookmarkMenu, newTabMenu]
-    let addTabMenuActionList = [addTabMenu]
-
-    toolbar?.tabsButton.menu = UIMenu(title: "", identifier: nil, children: tabButtonMenuActionList)
-    toolbar?.searchButton.menu = UIMenu(title: "", identifier: nil, children: addTabMenuActionList)
-    
-    topToolbar.tabsButton.menu = UIMenu(title: "", identifier: nil, children: tabButtonMenuActionList)
-    toolbar?.searchButton.menu = UIMenu(title: "", identifier: nil, children: addTabMenuActionList)
-    
-    // Update Actions for Add-Tab Button
-    toolbar?.addTabButton.menu = UIMenu(title: "", identifier: nil, children: addTabMenuActionList)
-  }
-}
-
-extension BrowserViewController: WKUIDelegate {
-  public func webView(_ webView: WKWebView, createWebViewWith configuration: WKWebViewConfiguration, for navigationAction: WKNavigationAction, windowFeatures: WKWindowFeatures) -> WKWebView? {
-    guard let parentTab = tabManager[webView] else { return nil }
-
-    guard !navigationAction.isInternalUnprivileged,
-          let navigationURL = navigationAction.request.url,
-          navigationURL.shouldRequestBeOpenedAsPopup()
-    else {
-      print("Denying popup from request: \(navigationAction.request)")
-      return nil
-    }
-
-    if let currentTab = tabManager.selectedTab {
-      screenshotHelper.takeScreenshot(currentTab)
-    }
-
-    // If the page uses `window.open()` or `[target="_blank"]`, open the page in a new tab.
-    // IMPORTANT!!: WebKit will perform the `URLRequest` automatically!! Attempting to do
-    // the request here manually leads to incorrect results!!
-    let newTab = tabManager.addPopupForParentTab(parentTab, configuration: configuration)
-
-    newTab.url = URL(string: "about:blank")
-    
-    toolbarVisibilityViewModel.toolbarState = .expanded
-
-    return newTab.webView
-  }
-
-  fileprivate func shouldDisplayJSAlertForWebView(_ webView: WKWebView) -> Bool {
-    // Only display a JS Alert if we are selected and there isn't anything being shown
-    return ((tabManager.selectedTab == nil ? false : tabManager.selectedTab!.webView == webView)) && (self.presentedViewController == nil)
-  }
-
-  public func webView(_ webView: WKWebView, runJavaScriptAlertPanelWithMessage message: String, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping () -> Void) {
-    var messageAlert = MessageAlert(message: message, frame: frame, completionHandler: completionHandler, suppressHandler: nil)
-    handleAlert(webView: webView, alert: &messageAlert) {
-      completionHandler()
-    }
-  }
-
-  public func webView(_ webView: WKWebView, runJavaScriptConfirmPanelWithMessage message: String, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping (Bool) -> Void) {
-    var confirmAlert = ConfirmPanelAlert(message: message, frame: frame, completionHandler: completionHandler, suppressHandler: nil)
-    handleAlert(webView: webView, alert: &confirmAlert) {
-      completionHandler(false)
-    }
-  }
-
-  public func webView(_ webView: WKWebView, runJavaScriptTextInputPanelWithPrompt prompt: String, defaultText: String?, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping (String?) -> Void) {
-    var textInputAlert = TextInputAlert(message: prompt, frame: frame, completionHandler: completionHandler, defaultText: defaultText, suppressHandler: nil)
-    handleAlert(webView: webView, alert: &textInputAlert) {
-      completionHandler(nil)
-    }
-  }
-
-  func suppressJSAlerts(webView: WKWebView) {
-    let script = """
-      window.alert=window.confirm=window.prompt=function(n){},
-      [].slice.apply(document.querySelectorAll('iframe')).forEach(function(n){if(n.contentWindow != window){n.contentWindow.alert=n.contentWindow.confirm=n.contentWindow.prompt=function(n){}}})
-      """
-    webView.evaluateSafeJavaScript(functionName: script, contentWorld: .defaultClient, asFunction: false)
-  }
-
-  func handleAlert<T: JSAlertInfo>(webView: WKWebView, alert: inout T, completionHandler: @escaping () -> Void) {
-    guard let promptingTab = tabManager[webView], !promptingTab.blockAllAlerts else {
-      suppressJSAlerts(webView: webView)
-      tabManager[webView]?.cancelQueuedAlerts()
-      completionHandler()
-      return
-    }
-    promptingTab.alertShownCount += 1
-    let suppressBlock: JSAlertInfo.SuppressHandler = { [unowned self] suppress in
-      if suppress {
-        func suppressDialogues(_: UIAlertAction) {
-          self.suppressJSAlerts(webView: webView)
-          promptingTab.blockAllAlerts = true
-          self.tabManager[webView]?.cancelQueuedAlerts()
-          completionHandler()
-        }
-        // Show confirm alert here.
-        let suppressSheet = UIAlertController(title: nil, message: Strings.suppressAlertsActionMessage, preferredStyle: .actionSheet)
-        suppressSheet.addAction(UIAlertAction(title: Strings.suppressAlertsActionTitle, style: .destructive, handler: suppressDialogues))
-        suppressSheet.addAction(
-          UIAlertAction(
-            title: Strings.cancelButtonTitle, style: .cancel,
-            handler: { _ in
-              completionHandler()
-            }))
-        if UIDevice.current.userInterfaceIdiom == .pad, let popoverController = suppressSheet.popoverPresentationController {
-          popoverController.sourceView = self.view
-          popoverController.sourceRect = CGRect(x: self.view.bounds.midX, y: self.view.bounds.midY, width: 0, height: 0)
-          popoverController.permittedArrowDirections = []
-        }
-        self.present(suppressSheet, animated: true)
-      } else {
-        completionHandler()
-      }
-    }
-    alert.suppressHandler = promptingTab.alertShownCount > 1 ? suppressBlock : nil
-    if shouldDisplayJSAlertForWebView(webView) {
-      let controller = alert.alertController()
-      controller.delegate = self
-      present(controller, animated: true)
-    } else {
-      promptingTab.queueJavascriptAlertPrompt(alert)
-    }
-  }
-
-  func checkIfWebContentProcessHasCrashed(_ webView: WKWebView, error: NSError) -> Bool {
-    if error.code == WKError.webContentProcessTerminated.rawValue && error.domain == "WebKitErrorDomain" {
-      print("WebContent process has crashed. Trying to reload to restart it.")
-      webView.reload()
-      return true
-    }
-
-    return false
-  }
-
-  public func webView(_ webView: WKWebView, contextMenuConfigurationForElement elementInfo: WKContextMenuElementInfo, completionHandler: @escaping (UIContextMenuConfiguration?) -> Void) {
-
-    // Only show context menu for valid links such as `http`, `https`, `data`. Safari does not show it for anything else.
-    // This is because you cannot open `javascript:something` URLs in a new page, or share it, or anything else.
-    guard let url = elementInfo.linkURL, url.isWebPage() else { return completionHandler(UIContextMenuConfiguration(identifier: nil, previewProvider: nil, actionProvider: nil)) }
-
-    let actionProvider: UIContextMenuActionProvider = { _ -> UIMenu? in
-      var actions = [UIAction]()
-
-      if let currentTab = self.tabManager.selectedTab {
-        let tabType = currentTab.type
-
-        if !tabType.isPrivate {
-          let openNewTabAction = UIAction(
-            title: Strings.openNewTabButtonTitle,
-            image: UIImage(systemName: "plus")
-          ) { _ in
-            self.addTab(url: url, inPrivateMode: false, currentTab: currentTab)
-          }
-
-          openNewTabAction.accessibilityLabel = "linkContextMenu.openInNewTab"
-
-          actions.append(openNewTabAction)
-        }
-
-        let openNewPrivateTabAction = UIAction(
-          title: Strings.openNewPrivateTabButtonTitle,
-          image: UIImage(named: "private_glasses", in: .module, compatibleWith: nil)!.template
-        ) { _ in
-          self.addTab(url: url, inPrivateMode: true, currentTab: currentTab)
-        }
-        openNewPrivateTabAction.accessibilityLabel = "linkContextMenu.openInNewPrivateTab"
-
-        actions.append(openNewPrivateTabAction)
-
-        let copyAction = UIAction(
-          title: Strings.copyLinkActionTitle,
-          image: UIImage(systemName: "doc.on.doc")
-        ) { _ in
-          UIPasteboard.general.url = url
-        }
-        copyAction.accessibilityLabel = "linkContextMenu.copyLink"
-
-        actions.append(copyAction)
-
-        if let braveWebView = webView as? BraveWebView {
-          let shareAction = UIAction(
-            title: Strings.shareLinkActionTitle,
-            image: UIImage(systemName: "square.and.arrow.up")
-          ) { _ in
-            let touchPoint = braveWebView.lastHitPoint
-            let touchRect = CGRect(origin: touchPoint, size: .zero)
-
-            // TODO: Find a way to add fixes #3323 and #2961 here:
-            // Normally we use `tab.temporaryDocument` for the downloaded file on the tab.
-            // `temporaryDocument` returns the downloaded file to disk on the current tab.
-            // Using a downloaded file url results in having functions like "Save to files" available.
-            // It also attaches the file (image, pdf, etc) and not the url to emails, slack, etc.
-            // Since this is **not** a tab but a standalone web view, the downloaded temporary file is **not** available.
-            // This results in the fixes for #3323 and #2961 not being included in this share scenario.
-            // This is not a regression, we simply never handled this scenario in both fixes.
-            // Some possibile fixes include:
-            // - Detect the file type and download it if necessary and don't rely on the `tab.temporaryDocument`.
-            // - Add custom "Save to file" functionality (needs investigation).
-            self.presentActivityViewController(
-              url, sourceView: braveWebView,
-              sourceRect: touchRect,
-              arrowDirection: .any)
-          }
-
-          shareAction.accessibilityLabel = "linkContextMenu.share"
-
-          actions.append(shareAction)
-        }
-
-        let linkPreview = Preferences.General.enableLinkPreview.value
-
-        let linkPreviewTitle = linkPreview ? Strings.hideLinkPreviewsActionTitle : Strings.showLinkPreviewsActionTitle
-        let linkPreviewAction = UIAction(title: linkPreviewTitle, image: UIImage(systemName: "eye.fill")) { _ in
-          Preferences.General.enableLinkPreview.value.toggle()
-        }
-
-        actions.append(linkPreviewAction)
-      }
-
-      return UIMenu(title: url.absoluteString.truncate(length: 100), children: actions)
-    }
-
-    let linkPreview: UIContextMenuContentPreviewProvider? = { [unowned self] in
-      if let tab = tabManager.tabForWebView(webView) {
-        return LinkPreviewViewController(url: url, for: tab, browserController: self)
-      }
-      return nil
-    }
-
-    let linkPreviewProvider = Preferences.General.enableLinkPreview.value ? linkPreview : nil
-    let config = UIContextMenuConfiguration(
-      identifier: nil, previewProvider: linkPreviewProvider,
-      actionProvider: actionProvider)
-
-    completionHandler(config)
-  }
-
-  public func webView(_ webView: WKWebView, contextMenuForElement elementInfo: WKContextMenuElementInfo, willCommitWithAnimator animator: UIContextMenuInteractionCommitAnimating) {
-    guard let url = elementInfo.linkURL else { return }
-    webView.load(URLRequest(url: url))
-  }
-
-  fileprivate func addTab(url: URL, inPrivateMode: Bool, currentTab: Tab) {
-    let tab = self.tabManager.addTab(URLRequest(url: url), afterTab: currentTab, isPrivate: inPrivateMode)
-    if inPrivateMode && !PrivateBrowsingManager.shared.isPrivateBrowsing {
-      self.tabManager.selectTab(tab)
-    } else {
-      // We're not showing the top tabs; show a toast to quick switch to the fresh new tab.
-      let toast = ButtonToast(
-        labelText: Strings.contextMenuButtonToastNewTabOpenedLabelText, buttonText: Strings.contextMenuButtonToastNewTabOpenedButtonText,
-        completion: { buttonPressed in
-          if buttonPressed {
-            self.tabManager.selectTab(tab)
-          }
-        })
-      self.show(toast: toast)
-    }
-    self.toolbarVisibilityViewModel.toolbarState = .expanded
   }
 }
